+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Increased number of blocks used in land specific structure block palettes
 - Land Type Extensions now support structure sets.
 - New Cathedral Structure on the Battlefield
+- Added tooltip to rungs which have gated conditions
 
 ### Changed
 
@@ -36,6 +37,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Removed
 
 - Removed `minestuck:non_mirrored` recipe type
+- Removed `preEntryRungLimit` server config
 
 ### Contributors for this release